add comment count exception

+++ src/main/java/cn/cncc/community/community/controller/PublishController.java
@@ -19,7 +19,7 @@ public class PublishController
   @Autowired private QuestionService questionService;
 
   @GetMapping("/publish/{id}")
-  public String edit(@PathVariable(name = "id") Integer id, Model model)
+  public String edit(@PathVariable(name = "id") Long id, Model model)
   {
     QuestionDTO question = questionService.getById(id);
     model.addAttribute("title", question.getTitle());
@@ -38,7 +38,7 @@ public class PublishController
       @RequestParam(value = "title", required = false) String title,
       @RequestParam(value = "description", required = false) String description,
       @RequestParam(value = "tag", required = false) String tag,
-      @RequestParam(value = "id", required = false) Integer id,
+      @RequestParam(value = "id", required = false) Long id,
       HttpServletRequest request,
       Model model) {
     model.addAttribute("title", title);
@@ -70,7 +70,7 @@ public class PublishController
     question.setTitle(title);
     question.setDescription(description);
     question.setTag(tag);
-    question.setCreator(user.getId().intValue());
+    question.setCreator(user.getId());
     question.setId(id);
     questionService.createOrUpdate(question);
